requirements.txt
gensim
numpy
scikit-learn
pandas
python-dotenv
openai
nltk
Flask
